feat: select warehouse before instructions
