+++ packages/frontend/src/components/common/LinkButton.tsx
@@ -3,11 +3,19 @@ import React, { FC } from 'react';
 import { useHistory } from 'react-router-dom';
 import { EventData, useTracking } from '../../providers/TrackingProvider';
 
-const LinkButton: FC<
-    { href: string; trackingEvent?: EventData } & React.ComponentProps<
-        typeof AnchorButton
-    >
-> = ({ href, target, trackingEvent, ...rest }) => {
+interface LinkButtonProps extends React.ComponentProps<typeof AnchorButton> {
+    href: string;
+    trackingEvent?: EventData;
+    replace?: boolean;
+}
+
+const LinkButton: FC<LinkButtonProps> = ({
+    replace,
+    href,
+    target,
+    trackingEvent,
+    ...rest
+}) => {
     const history = useHistory();
     const { track } = useTracking();
     return (
@@ -21,7 +29,11 @@ const LinkButton: FC<
                 }
                 if (target === '_blank') return;
                 e.preventDefault();
-                history.push(href);
+                if (replace) {
+                    history.replace(href);
+                } else {
+                    history.push(href);
+                }
             }}
         />
     );
